Move options package to ui/options
