+++ internal/backend/gs/config.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/restic/restic/internal/errors"
-	"github.com/restic/restic/internal/options"
+	"github.com/restic/restic/internal/ui/options"
 )
 
 // Config contains all configuration necessary to connect to a Google Cloud Storage
