+++ internal/backend/local/config.go
@@ -4,7 +4,7 @@ import (
 	"strings"
 
 	"github.com/restic/restic/internal/errors"
-	"github.com/restic/restic/internal/options"
+	"github.com/restic/restic/internal/ui/options"
 )
 
 // Config holds all information needed to open a local repository.
