+++ internal/backend/sftp/config.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"github.com/restic/restic/internal/errors"
-	"github.com/restic/restic/internal/options"
+	"github.com/restic/restic/internal/ui/options"
 )
 
 // Config collects all information required to connect to an sftp server.
